Silence some false-positive warnings about narrowing
Cast signed type to size_t to silence a -Wnarrowing warning, because in
these cases the value is known to be non-negative.

+++ source/FileSystem/FAT/FatFile.cpp
@@ -174,7 +174,7 @@ std::pair<int, size_t> FatFile::read(void* const buffer, const size_t size)
 	if (ret < 0)
 		return {ufatErrorToErrorCode(ret), {}};
 
-	return {{}, ret};
+	return {{}, static_cast<size_t>(ret)};
 }
 
 int FatFile::rewind()
@@ -285,7 +285,7 @@ std::pair<int, size_t> FatFile::write(const void* const buffer, const size_t siz
 	if (ret < 0)
 		return {ufatErrorToErrorCode(ret), {}};
 
-	return {{}, ret};
+	return {{}, static_cast<size_t>(ret)};
 }
 
 }	// namespace distortos
